refactor : 사다리 게임 참가자와 게임 결과 중복되는 부분 제거
